+++ src/view/components/ItemThumb.svelte
@@ -3,6 +3,7 @@
 
    export let item;
    export let maximize = false;
+   export let draggable=true;
 
    async function drag(e) {
       let type = "Item";
@@ -53,14 +54,14 @@
 
 {#if tooltip}
    <div
-      class="ui-w-full ui-h-full ui-border-none !ui-rounded-md ui-bg-contain ui-bg-no-repeat"
+      class="zoom-container ui-w-full ui-h-full ui-border-none !ui-rounded-md ui-bg-contain ui-bg-no-repeat"
       class:ui-w-8={!maximize}
       class:ui-h-8={!maximize}
       class:ui-bg-cover={maximize}
       style:background-image="url({thumb})"
       style:border={highlight}
       alt=""
-      draggable={true}
+    {draggable}
       on:pointerdown={() => null}
       on:click
       on:dragstart={drag}